Remove legacy implementation
