+++ ixp_tracker/conf.py
@@ -17,13 +17,6 @@ except AttributeError:
 except (TypeError, ValueError):
     raise ImproperlyConfigured("IXP_TRACKER_PEERING_DB_URL must be a string value")
 
-try:
-    IXP_TRACKER_ENABLE_EVENT_SOURCING = bool(settings.IXP_TRACKER_ENABLE_EVENT_SOURCING)
-except AttributeError:
-    IXP_TRACKER_ENABLE_EVENT_SOURCING = True
-except (TypeError, ValueError):
-    raise ImproperlyConfigured("IXP_TRACKER_ENABLE_EVENT_SOURCING must be a boolean")
-
 DATA_ARCHIVE_URL = "https://publicdata.caida.org/datasets/peeringdb/{year}/{month:02d}/peeringdb_2_dump_{year}_{month:02d}_{day:02d}.json"
 # Optional local data archive. When backfilling data the lib will look here first, before trying to get the data from CAIDA
 IXP_TRACKER_LOCAL_DATA_ARCHIVE_PATH: str | None
