+++ ixp_tracker/stats.py
@@ -2,19 +2,12 @@ import logging
 from datetime import datetime, timedelta, timezone
 from typing import Dict, Iterable, List, TypedDict, Union
 
-from django.db.models import Q
-from django.db.models.expressions import F
 from django_countries import countries
 
 from ixp_tracker.ixp_tracker import IXPTracker
 
 from ixp_tracker.importers import AdditionalDataSources, build_app
-from ixp_tracker.ixp_tracker_aggregates import is_ixp_active
 from ixp_tracker.models import (
-    IXP,
-    IXPMembershipRecord,
-    StatsPerCountryLegacy,
-    StatsPerIXPLegacy,
     StatsPerIXP,
     StatsPerCountry,
 )
@@ -34,24 +27,22 @@ class CountryStats(TypedDict):
 def generate_stats(
     lookup: AdditionalDataSources,
     stats_date: datetime | None = None,
-    disable_event_sourcing: bool = False,
 ):
     stats_date = stats_date or datetime.now(timezone.utc)
-    es_app = build_app(None, disable_event_sourcing)
-    do_generate_stats(lookup, stats_date, es_app)
+    es_app = build_app()
+    do_generate_stats(lookup, es_app, stats_date)
 
 
 def do_generate_stats(
     lookup: AdditionalDataSources,
+    es_app: IXPTracker,
     stats_date: datetime | None = None,
-    es_app: IXPTracker | None = None,
 ):
     stats_date = stats_date or datetime.now(timezone.utc)
     stats_date = stats_date.replace(day=1)
     date_now = datetime.now(timezone.utc)
     date_12_months_ago = stats_date.replace(year=(stats_date.year - 1))
     date_last_month = (stats_date - timedelta(days=1)).replace(day=1)
-    ixps = IXP.objects.filter(created__lte=stats_date).all()
     all_stats_per_country: Dict[str, CountryStats] = {}
     for code, _ in list(countries):
         all_stats_per_country[code] = {
@@ -62,317 +53,101 @@ def do_generate_stats(
             "member_and_customer_asns": set(),
             "total_capacity": 0,
         }
-    if es_app is None:
-        all_memberships = (
-            IXPMembershipRecord.objects.filter(
-                Q(start_date__lte=stats_date)
-                & (Q(end_date=None) | Q(end_date__gte=stats_date))
-            )
-            .values(
-                "member__ixp_id",
-                "member__id",
-                "speed",
-                "is_rs_peer",
-                "member__asn__number",
-            )
-            .order_by(F("end_date").desc(nulls_first=True))
-            .all()
+    # Ensure we load the state of all IXPs as they were on the stats date
+    ixps = es_app.get_all_ixps(stats_date)
+    for ixp in ixps:
+        # We always save the stats per IXP after their created date so we can track stats across time (e.g. if an IXP becomes inactive then active again)
+        isoc_id = es_app.find_isoc_id(ixp.id)
+        members = ixp.get_members()
+        member_asns = list(members.keys())
+        member_count = len(member_asns)
+        total_capacity = sum([m.port_speed for m in members.values()])
+        rs_peers = [m.is_rs_peer for m in members.values() if m.is_rs_peer]
+        rs_peering_rate = (len(rs_peers) / member_count) if member_count > 0 else 0
+        country_routed_asns = lookup.get_routed_asns_for_country(
+            ixp.country_code, stats_date
         )
-        memberships_last_month = (
-            IXPMembershipRecord.objects.filter(
-                Q(start_date__lte=date_last_month)
-                & (Q(end_date=None) | Q(end_date__gte=date_last_month))
-            )
-            .values(
-                "member__ixp_id",
-                "member__id",
-                "speed",
-                "is_rs_peer",
-                "member__asn__number",
-            )
-            .order_by(F("end_date").desc(nulls_first=True))
-            .all()
+        domestic_network_membership = calculate_local_asns_members_rate(
+            member_asns, country_routed_asns
         )
-        all_memberships_12_months_ago = (
-            IXPMembershipRecord.objects.filter(
-                Q(start_date__lte=date_12_months_ago)
-                & (Q(end_date=None) | Q(end_date__gte=date_12_months_ago))
-            )
-            .values(
-                "member__ixp_id",
-                "member__id",
-                "speed",
-                "is_rs_peer",
-                "member__asn__number",
-            )
-            .order_by(F("end_date").desc(nulls_first=True))
-            .all()
+        customer_asns = lookup.get_customer_asns(member_asns, stats_date)
+        domestic_network_coverage = calculate_local_asns_members_rate(
+            member_asns + customer_asns, country_routed_asns
         )
-        for ixp in ixps:
-            logger.debug("Calculating growth stats for IXP", extra={"ixp": ixp.id})
-            members = [
-                membership
-                for membership in all_memberships
-                if membership["member__ixp_id"] == ixp.id
-            ]
-            num_members_last_month = len(
-                [
-                    membership
-                    for membership in memberships_last_month
-                    if membership["member__ixp_id"] == ixp.id
-                ]
+        members_12_months_ago = ixp.get_members(as_at=date_12_months_ago)
+        member_asns_12_months_ago = members_12_months_ago.keys()
+        members_left_in_last_12_months = [
+            asn for asn in member_asns_12_months_ago if asn not in member_asns
+        ]
+        members_joined_in_last_12_months = [
+            asn for asn in member_asns if asn not in member_asns_12_months_ago
+        ]
+        members_last_month = ixp.get_members(as_at=date_last_month)
+        num_members_last_month = len(members_last_month.keys())
+        growth_members = member_count - num_members_last_month
+        StatsPerIXP.objects.update_or_create(
+            ixp=isoc_id,
+            stats_date=stats_date.date(),
+            defaults={
+                "members": member_count,
+                "capacity": (total_capacity / 1000),
+                "domestic_network_membership": domestic_network_membership,
+                "domestic_network_coverage": domestic_network_coverage,
+                "rs_peering_rate": rs_peering_rate,
+                "members_joined_last_12_months": len(members_joined_in_last_12_months),
+                "members_left_last_12_months": len(members_left_in_last_12_months),
+                "monthly_members_change": growth_members,
+                "monthly_members_change_percent": calculate_growth_members_percent(
+                    growth_members, num_members_last_month
+                ),
+                "last_generated": date_now,
+            },
+        )
+        if all_stats_per_country.get(ixp.country_code, None) is None:
+            logger.warning(
+                "IXP has possible invalid country",
+                extra={"ixp": isoc_id, "country": ixp.country_code},
             )
-            members_12_months_ago = [
-                membership
-                for membership in all_memberships_12_months_ago
-                if membership["member__ixp_id"] == ixp.id
-            ]
-            member_count = len(members)
-            capacity = 0
-            rs_peers_count = 0
-            members_counted = set()
-            for membership in members:
-                # There shouldn't be any duplicates but add this check just in case
-                if membership["member__id"] in members_counted:
-                    logger.warning(
-                        "Duplicate member found",
-                        extra={"member": membership["member__id"]},
-                    )
-                    member_count -= 1
-                    continue
-                members_counted.add(membership["member__id"])
-                capacity += membership["speed"]
-                if membership["is_rs_peer"]:
-                    rs_peers_count += 1
-            ixp_country = ixp.country_code
-            country_stats = all_stats_per_country.get(ixp_country)
-            if country_stats is None:
-                logger.warning("Country not found", extra={"country": ixp_country})
-                country_stats = {
-                    "ixp_count": 0,
-                    "all_asns": None,
-                    "routed_asns": None,
-                    "member_asns": set(),
-                    "member_and_customer_asns": set(),
-                    "total_capacity": 0,
-                }
-                all_stats_per_country[ixp_country] = country_stats
-            if country_stats.get("all_asns") is None:
-                all_stats_per_country[ixp_country]["all_asns"] = (
-                    lookup.get_asns_for_country(ixp_country, stats_date)
-                )
-            if country_stats.get("routed_asns") is None:
-                all_stats_per_country[ixp_country]["routed_asns"] = (
-                    lookup.get_routed_asns_for_country(ixp_country, stats_date)
-                )
-            member_asns = [membership["member__asn__number"] for membership in members]
-            member_asns_12_months_ago = [
-                membership["member__asn__number"]
-                for membership in members_12_months_ago
-            ]
-            members_left_in_last_12_months = [
-                asn for asn in member_asns_12_months_ago if asn not in member_asns
-            ]
-            members_joined_in_last_12_months = [
-                asn for asn in member_asns if asn not in member_asns_12_months_ago
-            ]
-            customer_asns = lookup.get_customer_asns(member_asns, stats_date)
-            members_and_customers = set(member_asns + customer_asns)
-            local_asns_members_rate = calculate_local_asns_members_rate(
-                member_asns,
-                all_stats_per_country[ixp_country]["all_asns"],  # type: ignore
+        if ixp.active_status:
+            all_stats_per_country[ixp.country_code]["ixp_count"] += 1
+            all_stats_per_country[ixp.country_code]["member_asns"] |= set(member_asns)
+            all_stats_per_country[ixp.country_code]["member_and_customer_asns"] |= set(
+                member_asns
             )
-            local_routed_asns_members_rate = calculate_local_asns_members_rate(
-                member_asns,
-                all_stats_per_country[ixp_country]["routed_asns"],  # type: ignore
+            all_stats_per_country[ixp.country_code]["member_and_customer_asns"] |= set(
+                customer_asns
             )
-            local_routed_asns_members_customers_rate = (
-                calculate_local_asns_members_rate(
-                    members_and_customers,
-                    all_stats_per_country[ixp_country]["routed_asns"],  # type: ignore
-                )
+            # We count capacity for all members, i.e. an ASN member at 2 IXPs will have capacity at each included in the sum
+            all_stats_per_country[ixp.country_code]["total_capacity"] += total_capacity
+    for code, _ in list(countries):
+        country_stats = all_stats_per_country[code]
+        if country_stats.get("all_asns") is None:
+            country_stats["all_asns"] = lookup.get_asns_for_country(code, stats_date)
+        if country_stats.get("routed_asns") is None:
+            country_stats["routed_asns"] = lookup.get_routed_asns_for_country(
+                code, stats_date
             )
-            rs_peering_rate = rs_peers_count / member_count if rs_peers_count else 0
-            growth_members = member_count - num_members_last_month
-            # We always save the stats per IXP so we can track stats across time (e.g. if an IXP becomes inactive then active again)
-            StatsPerIXPLegacy.objects.update_or_create(
-                ixp=ixp,
-                stats_date=stats_date.date(),
-                defaults={
-                    "ixp": ixp,
-                    "stats_date": stats_date.date(),
-                    "members": member_count,
-                    "capacity": (capacity / 1000),
-                    "local_asns_members_rate": local_asns_members_rate,
-                    "local_routed_asns_members_rate": local_routed_asns_members_rate,
-                    "local_routed_asns_members_customers_rate": local_routed_asns_members_customers_rate,
-                    "rs_peering_rate": rs_peering_rate,
-                    "members_joined_last_12_months": len(
-                        members_joined_in_last_12_months
-                    ),
-                    "members_left_last_12_months": len(members_left_in_last_12_months),
-                    "monthly_members_change": growth_members,
-                    "monthly_members_change_percent": calculate_growth_members_percent(
-                        growth_members, num_members_last_month
-                    ),
-                    "last_generated": date_now,
-                },
-            )
-            # Only aggregate this IXP's stats into the country stats if it's active
-            if is_ixp_active(members):
-                all_stats_per_country[ixp_country]["ixp_count"] += 1
-                # We only count unique ASNs that are members of an IXP in a country
-                all_stats_per_country[ixp_country]["member_asns"] |= set(member_asns)
-                all_stats_per_country[ixp_country]["member_and_customer_asns"] |= (
-                    members_and_customers
-                )
-                # But we count capacity for all members, i.e. an ASN member at 2 IXPs will have capacity at each included in the sum
-                all_stats_per_country[ixp_country]["total_capacity"] += capacity
-        for code, _ in list(countries):
-            country_stats = all_stats_per_country[code]
-            if country_stats.get("all_asns") is None:
-                country_stats["all_asns"] = lookup.get_asns_for_country(
-                    code, stats_date
-                )
-            if country_stats.get("routed_asns") is None:
-                country_stats["routed_asns"] = lookup.get_routed_asns_for_country(
-                    code, stats_date
-                )
-            local_asns_members_rate = calculate_local_asns_members_rate(
-                country_stats["member_asns"],
-                list(country_stats["all_asns"] or []),
-            )
-            local_routed_asns_members_rate = calculate_local_asns_members_rate(
-                country_stats["member_asns"],
-                list(country_stats["routed_asns"] or []),
-            )
-            local_routed_asns_members_customers_rate = (
-                calculate_local_asns_members_rate(
-                    country_stats["member_and_customer_asns"],
-                    list(country_stats["routed_asns"] or []),
-                )
-            )
-            StatsPerCountryLegacy.objects.update_or_create(
-                country_code=code,
-                stats_date=stats_date.date(),
-                defaults={
-                    "ixp_count": country_stats["ixp_count"],
-                    "asn_count": len(country_stats["all_asns"]),  # type: ignore
-                    "routed_asn_count": len(country_stats["routed_asns"]),  # type: ignore
-                    "member_count": len(country_stats["member_asns"]),
-                    "asns_ixp_member_rate": local_asns_members_rate,
-                    "routed_asns_ixp_member_rate": local_routed_asns_members_rate,
-                    "routed_asns_ixp_member_customers_rate": local_routed_asns_members_customers_rate,
-                    "total_capacity": (country_stats["total_capacity"] / 1000),
-                    "last_generated": date_now,
-                },
-            )
-    else:
-        # Ensure we load the state of all IXPs as they were on the stats date
-        ixps = es_app.get_all_ixps(stats_date)
-        for ixp in ixps:
-            # We always save the stats per IXP after their created date so we can track stats across time (e.g. if an IXP becomes inactive then active again)
-            isoc_id = es_app.find_isoc_id(ixp.id)
-            members = ixp.get_members()
-            member_asns = list(members.keys())
-            member_count = len(member_asns)
-            total_capacity = sum([m.port_speed for m in members.values()])
-            rs_peers = [m.is_rs_peer for m in members.values() if m.is_rs_peer]
-            rs_peering_rate = (len(rs_peers) / member_count) if member_count > 0 else 0
-            country_routed_asns = lookup.get_routed_asns_for_country(
-                ixp.country_code, stats_date
-            )
-            domestic_network_membership = calculate_local_asns_members_rate(
-                member_asns, country_routed_asns
-            )
-            customer_asns = lookup.get_customer_asns(member_asns, stats_date)
-            domestic_network_coverage = calculate_local_asns_members_rate(
-                member_asns + customer_asns, country_routed_asns
-            )
-            members_12_months_ago = ixp.get_members(as_at=date_12_months_ago)
-            member_asns_12_months_ago = members_12_months_ago.keys()
-            members_left_in_last_12_months = [
-                asn for asn in member_asns_12_months_ago if asn not in member_asns
-            ]
-            members_joined_in_last_12_months = [
-                asn for asn in member_asns if asn not in member_asns_12_months_ago
-            ]
-            members_last_month = ixp.get_members(as_at=date_last_month)
-            num_members_last_month = len(members_last_month.keys())
-            growth_members = member_count - num_members_last_month
-            StatsPerIXP.objects.update_or_create(
-                ixp=isoc_id,
-                stats_date=stats_date.date(),
-                defaults={
-                    "members": member_count,
-                    "capacity": (total_capacity / 1000),
-                    "domestic_network_membership": domestic_network_membership,
-                    "domestic_network_coverage": domestic_network_coverage,
-                    "rs_peering_rate": rs_peering_rate,
-                    "members_joined_last_12_months": len(
-                        members_joined_in_last_12_months
-                    ),
-                    "members_left_last_12_months": len(members_left_in_last_12_months),
-                    "monthly_members_change": growth_members,
-                    "monthly_members_change_percent": calculate_growth_members_percent(
-                        growth_members, num_members_last_month
-                    ),
-                    "last_generated": date_now,
-                },
-            )
-            if all_stats_per_country.get(ixp.country_code, None) is None:
-                logger.warning(
-                    "IXP has possible invalid country",
-                    extra={"ixp": isoc_id, "country": ixp.country_code},
-                )
-            if ixp.active_status:
-                all_stats_per_country[ixp.country_code]["ixp_count"] += 1
-                all_stats_per_country[ixp.country_code]["member_asns"] |= set(
-                    member_asns
-                )
-                all_stats_per_country[ixp.country_code]["member_and_customer_asns"] |= (
-                    set(member_asns)
-                )
-                all_stats_per_country[ixp.country_code]["member_and_customer_asns"] |= (
-                    set(customer_asns)
-                )
-                # We count capacity for all members, i.e. an ASN member at 2 IXPs will have capacity at each included in the sum
-                all_stats_per_country[ixp.country_code]["total_capacity"] += (
-                    total_capacity
-                )
-        for code, _ in list(countries):
-            country_stats = all_stats_per_country[code]
-            if country_stats.get("all_asns") is None:
-                country_stats["all_asns"] = lookup.get_asns_for_country(
-                    code, stats_date
-                )
-            if country_stats.get("routed_asns") is None:
-                country_stats["routed_asns"] = lookup.get_routed_asns_for_country(
-                    code, stats_date
-                )
-            local_routed_asns_members_rate = calculate_local_asns_members_rate(
-                country_stats["member_asns"],
-                list(country_stats["routed_asns"] or []),
-            )
-            local_routed_asns_members_customers_rate = (
-                calculate_local_asns_members_rate(
-                    country_stats["member_and_customer_asns"],
-                    list(country_stats["routed_asns"] or []),
-                )
-            )
-            StatsPerCountry.objects.update_or_create(
-                country_code=code,
-                stats_date=stats_date.date(),
-                defaults={
-                    "ixp_count": country_stats["ixp_count"],
-                    "routed_asn_count": len(country_stats["routed_asns"] or []),
-                    "member_count": len(country_stats["member_asns"]),
-                    "domestic_network_membership": local_routed_asns_members_rate,
-                    "domestic_network_coverage": local_routed_asns_members_customers_rate,
-                    "total_capacity": (country_stats["total_capacity"] / 1000),
-                    "last_generated": date_now,
-                },
-            )
+        local_routed_asns_members_rate = calculate_local_asns_members_rate(
+            country_stats["member_asns"],
+            list(country_stats["routed_asns"] or []),
+        )
+        local_routed_asns_members_customers_rate = calculate_local_asns_members_rate(
+            country_stats["member_and_customer_asns"],
+            list(country_stats["routed_asns"] or []),
+        )
+        StatsPerCountry.objects.update_or_create(
+            country_code=code,
+            stats_date=stats_date.date(),
+            defaults={
+                "ixp_count": country_stats["ixp_count"],
+                "routed_asn_count": len(country_stats["routed_asns"] or []),
+                "member_count": len(country_stats["member_asns"]),
+                "domestic_network_membership": local_routed_asns_members_rate,
+                "domestic_network_coverage": local_routed_asns_members_customers_rate,
+                "total_capacity": (country_stats["total_capacity"] / 1000),
+                "last_generated": date_now,
+            },
+        )
 
 
 def calculate_growth_members_percent(
