+++ tests/test_ixps_import.py
@@ -3,57 +3,80 @@ from datetime import datetime, timezone
 import pytest
 
 from ixp_tracker.importers import process_ixp_data
-from ixp_tracker.models import IXP
-from tests.fixtures import IXPFactory, MockLookup, PeeringIXFactory
+from tests.fixtures import MockLookup, PeeringIXFactory, build_app
 
 pytestmark = pytest.mark.django_db
+IXP_TRACKER_ENABLE_EVENT_SOURCING = True
 
 
 def test_with_no_data_does_nothing():
-    process_ixp_data([], datetime.now(timezone.utc), MockLookup())
+    app, _ = build_app()
+    process_ixp_data([], datetime.now(timezone.utc), MockLookup(), app)
 
-    ixps = IXP.objects.all()
-    assert len(ixps) == 0
+    assert len(app.get_all_ixps()) == 0
 
 
 def test_imports_a_new_ixp():
-    process_ixp_data([PeeringIXFactory()], datetime.now(timezone.utc), MockLookup())
+    app, _ = build_app()
+    process_ixp_data(
+        [PeeringIXFactory()], datetime.now(timezone.utc), MockLookup(), app
+    )
 
-    ixps = IXP.objects.all()
+    ixps = app.get_all_ixps()
     assert len(ixps) == 1
-    ixp = ixps.first()
-    assert not ixp.manrs_participant
 
 
-def test_updates_an_existing_ixp():
+def test_import_handles_missing_data():
     new_data = PeeringIXFactory()
-    manrs_participants = [new_data["id"]]
-    IXPFactory(
-        peeringdb_id=new_data["id"],
-        last_active=datetime(year=2024, month=4, day=1).replace(tzinfo=timezone.utc),
-        manrs_participant=False,
+    del new_data["fac_count"]
+    app, _ = build_app()
+    process_ixp_data([new_data], datetime.now(timezone.utc), MockLookup(), app)
+
+    ixps = app.get_all_ixps()
+    assert len(ixps) == 1
+
+
+def test_updates_an_existing_ixp(faker):
+    new_data = PeeringIXFactory()
+    app, _ = build_app()
+    city = faker.city()
+    name = f"{city} - IX"
+    long_name = f"{city} Internet Exchange Point"
+    app.import_ixp(
+        name,
+        long_name,
+        city,
+        new_data["id"],
+        faker.url(schemes=["https"]),
+        faker.country_code(),
+        faker.date_time_between(start_date="-1d", tzinfo=timezone.utc),
+        faker.date_time_between(start_date="-1d", tzinfo=timezone.utc),
+        datetime(year=2024, month=4, day=1).replace(tzinfo=timezone.utc, microsecond=1),
+        False,
+        False,
+        faker.random_number(digits=3),
+        faker.random_number(digits=2),
     )
 
     process_ixp_data(
         [new_data],
         datetime.now(timezone.utc),
-        MockLookup(manrs_participants=manrs_participants),
+        MockLookup(),
+        app,
     )
 
-    ixps = IXP.objects.all()
-    assert len(ixps) == 1
-    ixp = IXP.objects.first()
+    ixp = app.find_by_peeringdb_id(new_data["id"])
     assert ixp.last_active.date() == datetime.now(timezone.utc).date()
     assert ixp.name == new_data["name"]
-    assert ixp.manrs_participant
 
 
 def test_does_not_import_an_ixp_from_a_non_iso_country():
     new_data = PeeringIXFactory()
     new_data["country"] = "XK"  # XK is Kosovo, but it's not an official ISO code
-    process_ixp_data([new_data], datetime.now(timezone.utc), MockLookup())
+    app, _ = build_app()
+    process_ixp_data([new_data], datetime.now(timezone.utc), MockLookup(), app)
 
-    ixps = IXP.objects.all()
+    ixps = app.get_all_ixps()
     assert len(ixps) == 0
 
 
@@ -61,31 +84,40 @@ def test_handles_errors_with_source_data():
     data_with_problems = PeeringIXFactory()
     data_with_problems["created"] = "abc"
 
-    process_ixp_data([data_with_problems], datetime.now(timezone.utc), MockLookup())
+    app, _ = build_app()
+    process_ixp_data(
+        [data_with_problems], datetime.now(timezone.utc), MockLookup(), app
+    )
 
-    ixps = IXP.objects.all()
+    ixps = app.get_all_ixps()
     assert len(ixps) == 0
 
 
 def test_saves_manrs_participant():
     new_data = PeeringIXFactory()
     manrs_participants = [new_data["id"]]
+    app, _ = build_app()
     process_ixp_data(
         [new_data],
         datetime.now(timezone.utc),
         MockLookup(manrs_participants=manrs_participants),
+        app,
     )
 
-    ixp = IXP.objects.first()
+    ixp = app.find_by_peeringdb_id(new_data["id"])
     assert ixp.manrs_participant
 
 
 def test_saves_anchor_host():
     new_data = PeeringIXFactory()
     anchor_hosts = [new_data["id"]]
+    app, _ = build_app()
     process_ixp_data(
-        [new_data], datetime.now(timezone.utc), MockLookup(anchor_hosts=anchor_hosts)
+        [new_data],
+        datetime.now(timezone.utc),
+        MockLookup(anchor_hosts=anchor_hosts),
+        app,
     )
 
-    ixp = IXP.objects.first()
+    ixp = app.find_by_peeringdb_id(new_data["id"])
     assert ixp.anchor_host
